requirements.txt
yfinance==0.2.18
pandas==2.0.2
numpy==1.24.3
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
textblob==0.17.1
scikit-learn==1.2.2
